requirements-postgresql.txt
Flask
python-dotenv
requests
sentry-sdk
cryptography==39.0.2
argon2-cffi
CustomModules[loghandler,bitmaphandler]
aiosqlite
aiomysql
asyncpg
psycopg[binary,pool]
motor
